Fix connection handle data type
conn_handle should be uint16_t, not uint8_t.

Test: compilation test
Bug: 64232952
Merged-In: Ibce88e2cf2f74f402ea26f7471e5ac35aef6229b

+++ stack/btm/btm_acl.cc
@@ -293,7 +293,7 @@ void btm_acl_created(const RawAddress& bda, DEV_CLASS dc, BD_NAME bdn,
   }
 }
 
-void btm_acl_update_conn_addr(uint8_t conn_handle, const RawAddress& address) {
+void btm_acl_update_conn_addr(uint16_t conn_handle, const RawAddress& address) {
   uint8_t idx = btm_handle_to_acl_index(conn_handle);
   if (idx != MAX_L2CAP_LINKS) {
     btm_cb.acl_db[idx].conn_addr = address;
